+++ fwoa-core/authz-smart/src/smartScopeHelper.ts
@@ -117,7 +117,8 @@ function isSmartScopeSufficientForBulkDataAccess(
     if (bulkDataAuth.exportType === 'system') {
       bulkDataRequestHasCorrectScope = hasSufficientScopeType && resourceType === '*' && hasReadPermissions;
     } else if (bulkDataAuth.exportType === 'group') {
-      bulkDataRequestHasCorrectScope = ['system'].includes(scopeType) && hasReadPermissions;
+      bulkDataRequestHasCorrectScope =
+        ['system'].includes(scopeType) && hasReadPermissions && resourceType === '*';
     }
     return bulkDataRequestHasCorrectScope;
   }
@@ -202,4 +203,14 @@ export function filterOutUnusableScope(
   );
 
   return filteredScopes;
+}
+
+export function validateTokenScopes(scopes: string[], patientContext?: string, fhirUser?: string): void {
+  rejectInvalidScopeCombination(scopes);
+  if (scopes.some((scope: string) => scope.startsWith('patient/')) && !patientContext) {
+    throw new UnauthorizedError('Invalid patient scopes in token.');
+  }
+  if (scopes.some((scope: string) => scope.startsWith('user/')) && !fhirUser) {
+    throw new UnauthorizedError('Invalid user scopes in token.');
+  }
 }