+++ solutions/deployment/lib/kms.ts
@@ -1,6 +1,7 @@
 import { AccountRootPrincipal, Effect, PolicyDocument, PolicyStatement, ServicePrincipal } from 'aws-cdk-lib/aws-iam';
 import { Alias, Key } from 'aws-cdk-lib/aws-kms';
 import { Construct } from 'constructs';
+import { RemovalPolicy } from 'aws-cdk-lib';
 
 export default class KMSResources {
     backupKMSKey: Key;
@@ -15,6 +16,8 @@ export default class KMSResources {
 
     snsKMSKey: Key;
 
+    securityLogKMSKey?: Key;
+
     s3Alias: Alias;
 
     dynamoDbAlias: Alias;
@@ -25,7 +28,9 @@ export default class KMSResources {
 
     snsAlias: Alias;
 
-    constructor(scope: Construct, region: string, stage: string, account: string) {
+    securityLogAlias?: Alias;
+
+    constructor(scope: Construct, region: string, stage: string, account: string, enableSecurityLogging: boolean) {
         this.backupKMSKey = new Key(scope, 'backupKMSKey', {
             description: 'Encryption key for daily',
             enableKeyRotation: true,
@@ -84,7 +89,7 @@ export default class KMSResources {
                         effect: Effect.ALLOW,
                         actions: ['kms:*'],
                         resources: ['*'],
-                        principals: [new AccountRootPrincipal()],
+                        principals: [new AccountRootPrincipal(), new ServicePrincipal(`logs.${region}.amazonaws.com`)],
                     }),
                 ],
             }),
@@ -124,6 +129,29 @@ export default class KMSResources {
             }),
         });
 
+        if (enableSecurityLogging) {
+            this.securityLogKMSKey = new Key(scope, 'securityLogKMSKey', {
+                enableKeyRotation: true,
+                description: 'KMS CMK for Security Logger Middleware',
+                policy: new PolicyDocument({
+                    statements: [
+                        new PolicyStatement({
+                            sid: 'Allow root principal to manager key',
+                            effect: Effect.ALLOW,
+                            actions: ['kms:*'],
+                            resources: ['*'],
+                            principals: [new AccountRootPrincipal()],
+                        }),
+                    ],
+                }),
+            });
+
+            this.securityLogKMSKey.applyRemovalPolicy(RemovalPolicy.RETAIN);
+            this.securityLogAlias = new Alias(scope, 'securityLogAlias', {
+                aliasName: `alias/securityLogKMSKey-${stage}`,
+                targetKey: this.securityLogKMSKey,
+            });
+        }
         this.snsKMSKey = new Key(scope, 'snsKMSKey', {
             enableKeyRotation: true,
             description: 'KMS CMK for SNS',