+++ solutions/smart-deployment/src/lib/cdk-infra.test.ts
@@ -21,6 +21,10 @@ describe('cdk-nag AwsSolutions Pack', () => {
   const oAuth2ApiEndpoint: string = 'test';
   const patientPickerEndpoint: string = 'test';
   const validateXHTML: boolean = false;
+  const igMemoryLimit: number = 128;
+  const igMemorySize: number = 2048;
+  const igStorageSize: number = 512;
+  const enableSecurityLogging: boolean = true;
 
   beforeAll(() => {
     // GIVEN
@@ -47,8 +51,12 @@ describe('cdk-nag AwsSolutions Pack', () => {
       enableBackup,
       fhirVersion,
       validateXHTML,
+      igMemoryLimit,
+      igMemorySize,
+      igStorageSize,
       description:
-        'Test - Primary Template - This template creates all the necessary resources to deploy FHIR Works on AWS; a framework to deploy a FHIR server on AWS. It will be used to run cdk-nag.'
+        'Test - Primary Template - This template creates all the necessary resources to deploy FHIR Works on AWS; a framework to deploy a FHIR server on AWS. It will be used to run cdk-nag.',
+      enableSecurityLogging
     });
 
     fs.rm('./pnpm-lock.yaml', { force: true }, () => {});